Fix audio playback on Linux
On Linux, audio playback was slow and choppy. Profiling revealed
resampling was taking a lot of time, so that was switched to a
synchronous `FixedIn` resampler, which is a lot faster than the
previous one. In addition, many memory allocations were eliminated
by reusing source buffers and doing as much as possible in-place.

Further optimization is possible, but this paired with increasing
the buffer size has eliminated the choppiness. The flamegraph now
shows the UI taking more CPU time than the player thread.

There are still other significant issues on Linux after this change.