Experiment with getting millenium-player running in Flatpak
This was a hack to see if millenium-player works in the Flatpak
org.gnome.Platform runtime. It works by copying an already compiled
millenium-player release binary into the Flatpak image with
appropriate permissions.

With millenium-player compiled on Pop!_OS, the resulting binary
does successfully run in Flatpak, but it runs into KMS permissions
issues when attempting initialize WebGL:

```
KMS: DRM_IOCTL_MODE_CREATE_DUMB failed: Permission denied
```

I had to use an older, unsupported, version of the Gnome platform
since the current one at time of writing, 45, seems to be incompatible.
In 45, the entire window fails to open due to a similar error as seen
above.

Given that WebGL doesn't work in the runtime, it is likely that
WebkitGTK must be compiled from scratch with the Flatpak builder,
and if that's going to be done, then the entire application may as
well be compiled in it.

That's a problem for another day though.